Relocate errors from mod.rs to errors.rs
